CavityFlow version 5

+++ src/glirulus.d/glirulus.hpp
@@ -9,6 +9,6 @@ double L_inf(vector<double>&x);
 int glirulus_init(int argc, char **argv);
 int glirulus_mm(sparse::matrix<double>&A,vector<double>&b);
 vector<double> glirulus(sparse::matrix<double>&A,vector<double>&b);
-int glirulus_check(sparse::matrix<double>&A,vector<double>&x,vector<double>&b);
+double glirulus_check(sparse::matrix<double>&A,vector<double>&x,vector<double>&b);
 
 #endif
